+++ examples/custom_exception_ruby/lib/ahriman/error.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+module Ahriman
+  class Error < StandardError; end
+  class RubricError < Error; end
+end