+++ examples/custom_exception_ruby/test/error_test.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+require "test/unit"
+require_relative "../lib/ahriman"
+
+class ErrorTest < Test::Unit::TestCase
+
+  def test_error?
+    assert_raise(Ahriman::RubricError) { Ahriman::cast_rubric }
+  end
+
+end