+++ src/r_struct.rs
@@ -257,9 +257,8 @@ unsafe impl value::private::ReprValue for RStruct {
 }
 
 impl TryConvert for RStruct {
-    #[inline]
-    fn try_convert(val: &Value) -> Result<Self, Error> {
-        Self::from_value(*val).ok_or_else(|| {
+    fn try_convert(val: Value) -> Result<Self, Error> {
+        Self::from_value(val).ok_or_else(|| {
             Error::new(
                 exception::type_error(),
                 format!("no implicit conversion of {} into Struct", unsafe {